move agent api definition to its own file, add username to sendMessage request
